+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## 1.5.0
 
+* Change `oce.plot.ts()` by adding `simplify` argument.
 * Fix `[[` to yield ITS90 temperature for all classes, not just `ctd`.
 * Fix `as.ctd()` handling of temperature scale of first argument.
 * Fix `as.section()` handling of list of `argo` objects as first argument.